Move benchmark code to probe-rs-cli
