+++ cli/Cargo.toml
@@ -39,3 +39,7 @@ bitfield = "0.14.0"
 jep106 = "0.2.6"
 itm-decode = { version = "0.6", default-features = false }
 tracing-subscriber = { version = "0.3.16", features = ["env-filter"] }
+rand = "0.8.5"
+reqwest = { version = "0.11.13", features = ["blocking", "json"] }
+chrono = "0.4.23"
+serde = { version = "1.0.147", features = ["derive"] }
